// lib/generated/prisma/models/Announcement.ts
export type AnnouncementModel = runtime.Types.Result.DefaultSelection<Prisma.$AnnouncementPayload>
// ...
export type $AnnouncementPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Announcement"
  objects: {}
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    title: string
    summary: string
    category: string
    isPublished: boolean
    publishedAt: Date | null
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["announcement"]>
  composites: {}
}